Made modifications to the CSS layout, particularly the sidebar, and also moved my contact information to the sidebar.
In the current state of things, we had a fairly large sidebar
(280px wide). I decided to decrease the width of this sidebar
such that it was 200px wide instead. Consequently, I decreased the
font size of the large <h1>, the site title, from 60px to 36px.

I also decided to center the navigation elements because I thought
it looked nicer (the grid doesn't fit all cases...). Finally,
I moved my Twitter and Email contacts to the sidebar and added them
as buttons with rounded borders.