Modified my homepage content and also removed all references to the $scripts$ variable found on some of my templates.
I felt that my homepage was getting outdated and needed quite a bit
of an update. In addition, I had been meaning to overhaul this page,
since the language seemed a bit...childish.

I also made changes to the main hakyll.hs file to remove all
references to the $scripts$ variable because I no longer needed it
in my templates.