Rewrote the content on the projects.html page.
Not only was this page getting outdated, but it was also too long.
I made some changes to shorten each project description drastically.
I also added some projects to the list. At the moment, each project
listing is still a bit too long, so I'm planning to change some CSS
so that each list of links does not take up so much space.