Renamed email JS file, and fixed issue with classes not being rendered correctly in projects.html.
It made a lot more sense to call the email JS file as email.js, since
it is not limited strictly to home.html. I also added a little note
in home.html that makes it clear to those who disable JS or those
who use NoScript that my email is rendered with JS.

The class attributes in the output HTML were not being rendered
correctly in projects.html. This required replacing the pageCompiler
in hakyll.hs for projects.html with readPageCompiler.

Finally, I made changes to make all sublinks for each project to
appear as a single-line list of small links, as I felt that they
should not make such a large presence on the page.