Add CSRF support and change hidden button to hidden form
CSRF support: Add session and cookie support because CSRF requires it.
Modify /embed to generate a token to pass to user so it can be submitted
to /upload.

Hidden form update: we no longer attach .click handler to a button.
Instead, we attach a .submit handler that populates an input field in a
hidden form.

Finally, move upload code to its own file and add it as an export to
index.js.